Add timer to Game and tidy up readme.md
